SpecificBookPage, now shows book details and fixed an error in some Data Classes

+++ app/src/main/java/se/ju/bookapp/Android/Model/ListPriceX.kt
@@ -1,6 +1,6 @@
 package se.ju.bookapp.Android.Model
 
 data class ListPriceX(
-    val amountInMicros: Int,
+    val amountInMicros: Long,
     val currencyCode: String
 )
